requirements.txt
Django==3.1.6
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-simplejwt
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
django-role-permissions==3.1.1
drf-yasg==1.20.0
pymysql==1.0.2
django-filter==2.4.0